refactor(forms): merge the in/out Network and UoT duplicates
Both pairs edited the same field with the same option set — in/ built the
options from an array, out/ inlined them as <option> — and differed only in
that out/ hardcoded :mb="0" for the grid2 layouts it sits in.

Rather than pick a side and shift someone's spacing, mb becomes a prop
defaulting to 0. The nine out/protocols callers stay untouched; the five
standalone in/ callers pass :mb="15" to keep Field's usual gap. The in/ copies
are deleted and those callers now import ../out/.

Rendered the four affected in/ components and four out/ protocol forms and
hashed the markup against the same probe with this change stashed: all eight
byte identical, so neither side moved.
