+++ frontend/src/components/forms/in/TProxy.vue
@@ -1,13 +1,13 @@
 <template>
   <div>
     <SectionLabel style="margin-bottom: 12px;">TProxy</SectionLabel>
-    <Network :data="inbound" />
+    <Network :data="inbound" :mb="15" />
   </div>
 </template>
 
 <script lang="ts" setup>
 import SectionLabel from '@/components/ui/SectionLabel.vue'
-import Network from './Network.vue'
+import Network from '../out/Network.vue'
 
 defineProps<{ inbound: any }>()
 </script>
